Single Input Model, Bug Fixing, and Anaconda Environment Integration
Added a notebook for the single input version of the model. This version draws the bounding boxes of the objects in hot pink before training on the images. Fixed a bug where image values where normalized twice. Added "anaconda-project.yml" and "spec-file.txt" for easy duplication in Anaconda of the Environment.